completely changed tables to hash table; separated files; .y not yet refactored
